refactor: IFlowPage to common types
