+++ src/common/types.ts
@@ -1,0 +1,5 @@
+export interface IFlowPage {
+  name: string;
+  description: string;
+  component: React.ReactNode; // forms, summary. etc
+}
