+++ src/modules/CreateDrops/types/types.ts
@@ -1,9 +1,3 @@
-export interface IFlowPage {
-  name: string;
-  description: string;
-  component: React.ReactNode; // forms, summary. etc
-}
-
 export interface SummaryItem {
   type: 'text' | 'image';
   name: string;
