+++ src/pages/claim/ticket/[id].tsx
@@ -1,10 +1,10 @@
 import { Box } from '@chakra-ui/react';
 
 import { PageHead } from '@/common/components/PageHead';
+import { IFlowPage } from '@/common/types';
 
 import { ClaimFormContextProvider } from '@/modules/claim/ClaimFormContext';
 import { ClaimTicketFlowProvider } from '@/modules/claim/ticket/ClaimTicketFlowContext';
-import { IFlowPage } from '@/modules/CreateDrops/types/types';
 import { ClaimTicketFormFlow } from '@/modules/claim/ticket/ClaimTicketFormFlow';
 import { ClaimTicketSummaryFlow } from '@/modules/claim/ticket/ClaimTicketSummaryFlow';
 import { ClaimTicketFlow } from '@/modules/claim/ticket/ClaimTicketFlow';
